apps/api/requirements.txt
fastapi
strawberry-graphql
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose
alembic
pytest
pytest-asyncio
httpx
pytest-cov